requirements.txt
streamlit
pandas
scikit-learn
numpy
requests
groq
python-dotenv
